Refactor and add a test
